Plugins: categorize linux.keyboard_notifiers as a malware plugin
